Fully updated to 0.3.0 spec
