+++ beacon-java-core/README.md
@@ -1,0 +1,3 @@
+# Beacon Java Core
+
+A java module part of the Beacon Java Development Kit (BJDK) which contains some of the core functions used across serparate modules
