+++ beacon-java-core/src/main/resources/META-INF/beans.xml
