+++ beacon-java-rest/src/main/webapp/WEB-INF/jboss-web.xml
