+++ beacon-java-rest/src/main/webapp/index.jsp
