+++ beacon-java-rest/src/test/resources/arquillian.xml
