--- src/main/java/com/dnastack/beacon/providers/GsonMessageBodyHandler.java
@@ -1,110 +1,0 @@
-/*
- * The MIT License
- *
- * Copyright 2014 Miroslav Cupak (mirocupak@gmail.com).
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package com.dnastack.beacon.providers;
-
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-
-import javax.ws.rs.Consumes;
-import javax.ws.rs.Produces;
-import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.MultivaluedMap;
-import javax.ws.rs.ext.MessageBodyReader;
-import javax.ws.rs.ext.MessageBodyWriter;
-import javax.ws.rs.ext.Provider;
-import java.io.*;
-import java.lang.annotation.Annotation;
-import java.lang.reflect.Type;
-
-/**
- * Gson Message Body Handler for serializing JSON data to and from auto generated Java classes
- *
- * @author Patrick Magee    (patrickmageee@gmail.com)
- * @version 1.0
- */
-@Provider
-@Produces(MediaType.APPLICATION_JSON)
-@Consumes(MediaType.APPLICATION_JSON)
-public class GsonMessageBodyHandler implements MessageBodyReader<Object>, MessageBodyWriter<Object> {
-
-
-    private static final String UTF_8 = "UTF-8";
-
-    private Gson gson;
-
-    private Gson getGson() {
-        if (gson == null) {
-            gson = new GsonBuilder().create();
-        }
-        return gson;
-    }
-
-    @Override
-    public boolean isReadable(Class<?> aClass, Type type, Annotation[] annotations, MediaType mediaType) {
-        return true;
-    }
-
-    @Override
-    public Object readFrom(Class<Object> aClass, Type type, Annotation[] annotations, MediaType mediaType, MultivaluedMap<String, String> multivaluedMap, InputStream inputStream) throws IOException, WebApplicationException {
-        InputStreamReader streamReader = new InputStreamReader(inputStream, UTF_8);
-        try {
-            Type jsonType;
-            if (aClass.equals(type)) {
-                jsonType = aClass;
-            } else {
-                jsonType = type;
-            }
-            return getGson().fromJson(streamReader, jsonType);
-        } finally {
-            streamReader.close();
-        }
-    }
-
-    @Override
-    public boolean isWriteable(Class<?> aClass, Type type, Annotation[] annotations, MediaType mediaType) {
-        return true;
-    }
-
-    @Override
-    public long getSize(Object o, Class<?> aClass, Type type, Annotation[] annotations, MediaType mediaType) {
-        return -1;
-    }
-
-    @Override
-    public void writeTo(Object o, Class<?> aClass, Type type, Annotation[] annotations, MediaType mediaType, MultivaluedMap<String, Object> multivaluedMap, OutputStream outputStream) throws IOException, WebApplicationException {
-        OutputStreamWriter writer = new OutputStreamWriter(outputStream, UTF_8);
-        try {
-            Type jsonType;
-            if (type.equals(type)) {
-                jsonType = aClass;
-            } else {
-                jsonType = type;
-            }
-            getGson().toJson(o, jsonType, writer);
-        } finally {
-            writer.close();
-        }
-    }
-}
