--- src/main/java/com/dnastack/beacon/rest/BeaconInfo.java
@@ -1,60 +1,0 @@
-/*
- * The MIT License
- *
- * Copyright 2014 Miroslav Cupak (mirocupak@gmail.com).
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package com.dnastack.beacon.rest;
-
-import com.dnastack.beacon.service.BeaconService;
-import org.apache.avro.AvroRemoteException;
-import org.ga4gh.beacon.Beacon;
-
-import javax.inject.Inject;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-
-/**
- * Beacon rest resource.
- *
- * @author Miroslav Cupak (mirocupak@gmail.com)
- * @author Patrick Magee    (patrickmageee@gmail.com)
- * @version 1.0
- */
-@Path("/")
-@Produces({MediaType.APPLICATION_JSON})
-public class BeaconInfo {
-
-    @Inject
-    private BeaconService service;
-
-    /**
-     * Gets Information on the beacon
-     *
-     * @return Beacon represenation
-     * @throws AvroRemoteException
-     */
-    @GET
-    public Beacon info() throws AvroRemoteException {
-        return service.getBeacon();
-    }
-}
