--- src/main/resources/com/dnastack/beacon/entity/jaxb.properties
@@ -1,1 +1,0 @@
-javax.xml.bind.context.factory=org.eclipse.persistence.jaxb.JAXBContextFactory
