--- src/main/webapp/WEB-INF/web.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xmlns="http://java.sun.com/xml/ns/javaee"
-         xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd"
-         version="3.0">
-    <distributable/>
-
-    <welcome-file-list>
-        <welcome-file>index.jsp</welcome-file>
-    </welcome-file-list>
-
-    <context-param>
-        <param-name>resteasy.providers</param-name>
-        <param-value>com.dnastack.beacon.providers.GsonMessageBodyHandler</param-value>
-    </context-param>
-    <context-param>
-        <param-name>resteasy.use.builtin.providers</param-name>
-        <param-value>false</param-value>
-    </context-param>
-    
-    <session-config>
-        <session-timeout>
-            30
-        </session-timeout>
-    </session-config>
-</web-app>
